% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{app_cache_dir}
\alias{app_cache_dir}
\title{Path to an App Cache Directory}
\usage{
app_cache_dir(appname)
}
\arguments{
\item{appname}{Character; the name of the application that will "own" this
file, such as the name of a package.}
}
\value{
The full path to the app's cache directory.
}
\description{
App cache directories can depend on the user's operating system and an
overall \code{R_USER_CACHE_DIR} environment variable. This function
}
\examples{
app_cache_dir("myApp")
}
