% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{construct_processed_filename}
\alias{construct_processed_filename}
\title{Construct Processed Filename}
\usage{
construct_processed_filename(source_path, extension = "")
}
\arguments{
\item{source_path}{Character scalar; the full path to the source file.}

\item{extension}{Character scalar; an optional filename extension.}
}
\value{
A unique filename for a processed version of the file.
}
\description{
Given the path to a file, construct a unique filename using the hash of the
path.
}
\examples{
construct_processed_filename(
  source_path = "my/file.txt",
  extension = "rds"
)
}
