% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr-package.R
\docType{package}
\name{dlr-package}
\alias{dlr}
\alias{dlr-package}
\title{dlr: Download and Cache Files Safely}
\description{
The goal of dlr is to provide a friendly wrapper around the common pattern of downloading a file if that file does not already exist locally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/macmillancontentscience/dlr}
  \item Report bugs at \url{https://github.com/macmillancontentscience/dlr/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jonathan Bratt \email{jonathan.bratt@macmillan.com} (\href{https://orcid.org/0000-0003-2859-0076}{ORCID})
}

Other contributors:
\itemize{
  \item Bedford Freeman & Worth Pub Grp LLC DBA Macmillan Learning [copyright holder, funder]
}

}
\keyword{internal}
