% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{.cache_dir_messaging}
\alias{.cache_dir_messaging}
\title{Reusable Path Permissions Messaging}
\usage{
.cache_dir_messaging(appname, cache_dir, message_style = "abort", verbose)
}
\arguments{
\item{appname}{Character; the name of the application that will "own" the
cache, such as the name of a package.}

\item{cache_dir}{The directory that will be used for caching.}

\item{message_style}{Whether to "abort" (default) or issue a "warning".}

\item{verbose}{Whether to issue the messages (included for easier usage in
other functions).}
}
\value{
\code{TRUE} invisibly.
}
\description{
Reusable Path Permissions Messaging
}
\keyword{internal}
