% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{.do_function}
\alias{.do_function}
\title{Do a Function Call}
\usage{
.do_function(do_f, default_f, main_args, extra_args = NULL)
}
\arguments{
\item{do_f}{The function to call.}

\item{default_f}{The function to call if \code{do_f} is \code{NULL}.}

\item{main_args}{A list containing the first argument(s) to \code{do_f}.}

\item{extra_args}{A list containing any additional arguments to \code{do_f}}
}
\value{
The return from the call to \code{do_f}.
}
\description{
Do a Function Call
}
\keyword{internal}
