% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{.download_then}
\alias{.download_then}
\title{Compose a Download-then-Process Function}
\usage{
.download_then(process_f)
}
\arguments{
\item{process_f}{A function or one-sided formula to use to read and process
the raw file after it is downloaded. The path to the downloaded file will
be passed as the first argument to this function.}
}
\value{
The processed data returned by \code{process_f}.
}
\description{
Compose a Download-then-Process Function
}
\keyword{internal}
