% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{.read_file}
\alias{.read_file}
\title{Read/Process a File}
\usage{
.read_file(path, read_f, read_args)
}
\arguments{
\item{path}{Character scalar; the full path to the file.}

\item{read_f}{The function to use to read and process the file. The first
argument to this function must be the path to the file.}

\item{read_args}{Additional arguments to \code{read_f}.}
}
\value{
The processed file.
}
\description{
Read/Process a File
}
\keyword{internal}
