% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{.write_object}
\alias{.write_object}
\title{Write an Object}
\usage{
.write_object(object, target_path, write_f, write_args)
}
\arguments{
\item{object}{An R object.}

\item{target_path}{Character scalar; the full path to the file to write.}

\item{write_f}{The function to use to write the file. The first argument to
this function must be the object, and the second argument must be the path
to the file.}

\item{write_args}{Additional arguments to \code{write_f}.}
}
\value{
The return from the call to \code{write_f}.
}
\description{
Write an Object
}
\keyword{internal}
