% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_path}
\alias{download_path}
\title{Download to a Path}
\usage{
download_path(
  url,
  path,
  filename = fs::path_file(url),
  process_f = NULL,
  redownload = FALSE,
  ...
)
}
\arguments{
\item{url}{Character; the location of the file on the internet.}

\item{path}{Character; the local directory to which the file should be
written.}

\item{filename}{Character; an optional filename for the file (by default, the
filename of the remote file is used).}

\item{process_f}{A function or one-sided formula to load, process, and save
the downloaded files. The temporary path to the downloaded file is passed
as the first argument to this function, and the target path is passed as
the second argument. See the vignette for an example.}

\item{redownload}{Logical; should the file be redownloaded if it already
exists locally?}

\item{...}{Additional arguments passed to \code{process_f}.}
}
\value{
The full path to the local file.
}
\description{
This function downloads a file to a given path if that file does not already
exist. This is mostly a convenience wrapper around
\code{\link[utils]{download.file}}, with checks to make sure the path exists
and the file hasn't already been downloaded.
}
\examples{
\donttest{
if (interactive()) {
  download_path(
    url = "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co",
    path = tempdir()
  )

  # Or process before saving.
  csv_to_rds <- function(temp_file, final_file) {
    df <- read.csv(
      temp_file, header = TRUE, stringsAsFactors = FALSE
    )
    saveRDS(df, final_file)
  }

  download_path(
    url = "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co",
    path = tempdir(),
    filename = "austin_smoke_free.rds",
    process_f = csv_to_rds
  )
}
}
}
