% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grepvec.R
\name{grepvec}
\alias{grepvec}
\title{MySQL table: glove
Origianl data downloaded from glove website,glove.840B.300d_db.txt
Word filtering is case-sensitive.}
\usage{
grepvec(myword)
}
\arguments{
\item{myword:}{a word}
}
\value{
a vector represents the word
}
\description{
MySQL table: glove
Origianl data downloaded from glove website,glove.840B.300d_db.txt
Word filtering is case-sensitive.
}
\examples{
grepvec("the")
}

