% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{read_or_cache}
\alias{read_or_cache}
\title{Read or Cache a File}
\usage{
read_or_cache(
  source_path,
  appname,
  filename = construct_processed_filename(source_path),
  process_f = readRDS,
  process_args = NULL,
  read_f = readRDS,
  read_args = NULL,
  write_f = saveRDS,
  write_args = NULL,
  force_process = FALSE
)
}
\arguments{
\item{source_path}{Character scalar; the path to the raw file. Paths starting
with \code{http://}, \code{http://}, \code{http://}, or \code{http://} will
be downloaded to a temp file if the processed version is not already
available.}

\item{appname}{Character; the name of the application that will "own" the
cache, such as the name of a package.}

\item{filename}{Character; an optional filename for the cached version of the
file. By default, a filename is constructed using
\code{\link{construct_processed_filename}}}

\item{process_f}{A function or one-sided formula to use to process the source
file. \code{source_path} will be passed as the first argument to this
function. Defaults to \code{read_f}.}

\item{process_args}{An optional list of additional arguments to
\code{process_f}.}

\item{read_f}{A function or one-sided formula to use to read the processed
file. \code{target_path} will be passed as the first argument to this
function. Defaults to \code{\link[base]{readRDS}}.}

\item{read_args}{An optional list of additional arguments to \code{read_f}.}

\item{write_f}{A function or one-sided formula to use to save the processed
file. The processed object will be passed as the first argument to this
function, and \code{target_path} will be passed as the second argument.
Defaults to \code{\link[base]{saveRDS}}.}

\item{write_args}{An optional list of additional arguments to \code{write_f}.}

\item{force_process}{A logical scalar indicating whether we should process
the source file even if the target already exists. This can be particularly
useful if you wish to redownload a file.}
}
\value{
The processed object.
}
\description{
This function wraps \code{\link{read_or_process}}, specifying an app's cache
directory as the target directory.
}
\examples{
if (interactive()) {
  austin_smoke_free <- read_or_cache(
    "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co",
    appname = "dlr",
    process_f = read.csv
  )
  head(austin_smoke_free)
}

if (interactive()) {
  # Calling the function a second time gives the result instantly.
  austin_smoke_free <- read_or_process(
    "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co",
    appname = "dlr",
    process_f = read.csv
  )
  head(austin_smoke_free)
}

if (interactive()) {
  # Remove the generated file.
  unlink(
    construct_cache_filename(
      "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co"
    )
  )
}
}
