% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_functions.R
\name{set_timeout}
\alias{set_timeout}
\title{Set Download Timeout}
\usage{
set_timeout(seconds = 600L)
}
\arguments{
\item{seconds}{The number of seconds to set as the timeout (default 600
seconds).}
}
\value{
A list with the old \code{timeout} setting (invisibly).
}
\description{
The default timeout for downloads is 60 seconds. This is not long enough for
many of the files that are downloaded using this package. We therefore supply
a convenience function to easily change this setting. You can permanently
change this default by setting \code{R_DEFAULT_INTERNET_TIMEOUT} in your
\code{.Renviron}.
}
\examples{
getOption("timeout")
old_setting <- set_timeout()
getOption("timeout")
options(old_setting)
}
