% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{Create `fuego` tensor of arbitrary shapes}
\arguments{
\item{data}{numeric vector, array or data.frame}

\item{dims}{dimensions}

\item{requires.grad}{is gradient required; logical}
}
\description{
Tensor abstraction, an interface for CPU and CUDA tensors.
}
\examples{
library(dlr)
x   <- cpu_tensor(1:30, c(3, 10))
}
