% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr.R
\docType{package}
\name{dlr-package}
\alias{dlr}
\alias{dlr-package}
\title{dlr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
