% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_cache}
\alias{download_cache}
\title{Download to App Cache Dir}
\usage{
download_cache(
  url,
  appname,
  filename = fs::path_file(url),
  process_f = NULL,
  redownload = FALSE,
  ...
)
}
\arguments{
\item{url}{Character; the location of the file on the internet.}

\item{appname}{Character; the name of the application that will "own" this
file, such as the name of a package.}

\item{filename}{Character; an optional filename for the file (by default, the
filename of the remote file is used).}

\item{process_f}{A function or one-sided formula to load, process, and save
the downloaded files. The temporary path to the downloaded file is passed
as the first argument to this function, and the target path is passed as
the second argument. See the vignette for an example.}

\item{redownload}{Logical; should the file be redownloaded if it already
exists locally?}

\item{...}{Additional arguments passed to \code{process_f}.}
}
\value{
The full path to the local file.
}
\description{
Download files to an application cache dir, using platform conventions as
defined in \code{\link[rappdirs]{user_cache_dir}}.
}
\examples{
\donttest{
if (interactive()) {
  download_cache(
    url = "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co",
    appname = "dlr"
  )

  # Or process before saving.
  csv_to_rds <- function(temp_file, final_file) {
    df <- read.csv(
      temp_file, header = TRUE, stringsAsFactors = FALSE
    )
    saveRDS(df, final_file)
  }

  download_cache(
    url = "https://query.data.world/s/owqxojjiphaypjmlxldsp566lck7co",
    appname = "dlr",
    filename = "austin_smoke_free.rds",
    process_f = csv_to_rds
  )
}
}
}
